fix:relationship service 에서 유저를 중복 팔로우하는 문제

+++ src/main/java/com/swp/user/domain/RelationshipService.java
@@ -2,6 +2,7 @@ package com.swp.user.domain;
 
 import com.swp.auth.dto.JwtUserDetails;
 import com.swp.user.dto.RelationshipRequestDto;
+import com.swp.user.exception.RelationshipConflictException;
 import com.swp.user.exception.RelationshipNotFoundException;
 import com.swp.user.exception.UserNotFoundException;
 import lombok.RequiredArgsConstructor;
@@ -16,33 +17,42 @@ public class RelationshipService {
     private final UserRepository userRepository;
 
     @Transactional
-    public void createRelationship(RelationshipRequestDto relationshipRequestDto){
-        JwtUserDetails userDetails = (JwtUserDetails)SecurityContextHolder.getContext()
-            .getAuthentication()
-            .getPrincipal();
-        User fromUser=userRepository.findByProviderAndProviderId(userDetails.getProvider(), userDetails.getUsername())
-            .orElseThrow(() -> new UserNotFoundException("없는 유저입니다"));
-        User toUser= userRepository.findById(relationshipRequestDto.getToUserId())
-            .orElseThrow(() -> new UserNotFoundException("팔로우할 유저를 찾을 수 없습니다."));
-        relationshipRepository.save(Relationship.builder()
-            .fromUser(fromUser)
-            .toUser(toUser)
-            .build());
+    public void createRelationship(RelationshipRequestDto relationshipRequestDto) {
+        JwtUserDetails userDetails = (JwtUserDetails) SecurityContextHolder.getContext()
+                .getAuthentication()
+                .getPrincipal();
+        User fromUser = userRepository.findByProviderAndProviderId(userDetails.getProvider(), userDetails.getUsername())
+                .orElseThrow(() -> new UserNotFoundException("없는 유저입니다"));
+        User toUser = userRepository.findById(relationshipRequestDto.getToUserId())
+                .orElseThrow(() -> new UserNotFoundException("팔로우할 유저를 찾을 수 없습니다."));
+
+        if (relationshipRepository.existsByFromUserAndToUser(fromUser, toUser)) {
+            throw new RelationshipConflictException("이미 해당 유저를 팔로우하고 있습니다.");
+        } else {
+            relationshipRepository.save(Relationship.builder()
+                    .fromUser(fromUser)
+                    .toUser(toUser)
+                    .build());
+        }
     }
 
     @Transactional
-    public void deleteRelationship(RelationshipRequestDto relationshipRequestDto){
-        JwtUserDetails userDetails = (JwtUserDetails)SecurityContextHolder.getContext()
+    public void deleteRelationship(RelationshipRequestDto relationshipRequestDto) {
+        JwtUserDetails userDetails = (JwtUserDetails) SecurityContextHolder.getContext()
                 .getAuthentication()
                 .getPrincipal();
-        User fromUser=userRepository.findByProviderAndProviderId(userDetails.getProvider(), userDetails.getUsername())
+        User fromUser = userRepository.findByProviderAndProviderId(userDetails.getProvider(), userDetails.getUsername())
                 .orElseThrow(() -> new UserNotFoundException("없는 유저입니다"));
-        User toUser= userRepository.findById(relationshipRequestDto.getToUserId())
+        User toUser = userRepository.findById(relationshipRequestDto.getToUserId())
                 .orElseThrow(() -> new UserNotFoundException("팔로우 취소할 유저를 찾을 수 없습니다."));
 
-        relationshipRepository.delete(
-                relationshipRepository.findByFromUserAndToUser(fromUser, toUser)
-                        .orElseThrow(() -> new RelationshipNotFoundException(toUser.getNickname() + "유저를 팔로우하고 있지 않습니다."))
-        );
+        if (relationshipRepository.existsByFromUserAndToUser(fromUser, toUser)) {
+            relationshipRepository.delete(
+                    relationshipRepository.findByFromUserAndToUser(fromUser, toUser)
+                            .orElseThrow(() -> new RelationshipNotFoundException(toUser.getNickname() + "유저를 팔로우하고 있지 않습니다."))
+            );
+        } else {
+            throw new RelationshipNotFoundException("해당 유저를 팔로우하고 있지 있습니다.");
+        }
     }
 }
